chore: move loose parser into public oci package On-behalf-of: Gergely Brautigam <gergely.brautigam@sap.com>
